delete file when delete product

+++ controllers/admin.js
@@ -1,6 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require("express-validator/check")
-
+const fileHelper = require("./../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -143,6 +143,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save()
@@ -176,12 +177,19 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  // Product.findByIdAndRemove(prodId)
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('product not found'))
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(result => {
       res.redirect("/admin/products")
     })
     .catch(err => {
+      console.log(err)
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error)
